critical vulnerabilities & outdated code
The code is outdated, electron has evolved and no longer supports certain functions needed for this project. it was a fun learning experience. read the README if you want more information. or look over my commits for some fixes that i was able to locate on the creators video comment feed. overall was fun to try this out.